refactor(web): stack fallback model fields vertically + advanced spacing
Extract the fallback-model row into its own component and lay it out like the main
agent form — provider on top, model below, credential under it (gated on
own-keys) — inside a labeled per-fallback card, instead of a cramped single line.
Add micro-spacing between labels and inputs in the advanced settings block.
